requirements.txt
Flask==3.1.2
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.31.0